Add script to create 2016 fixtures. Add environ.
Created a new fixtures file in data/fixtures/nfl-schedule-2016.tsv. Also
updated the rosters and put them in a file name which includes the date.

Added environ so that we can configure the logging directory via an
environment variable. Next, I'm going to attempt to upgrade the log4j
version so that we can reference environment variables directly in the
log4j.properties file.

Tested like this:

  lein with-profile dev test