Resurrection for 2016. Catch exceptions from SNS.
Needed to explicitly load the fixtures for 2016.

Added a try-catch around the post to SNS - otherwise the main thread
just stops when we get any exception from AWS.

Also a bug fix related to the setup of the CSV reporter for our metrics.
The reporter was def'ed, so the environment variable that defines the
log location was resolved at compile time. I wrapped it in a defer so
it's resolved lazily.